+++ ai_snake_lab/game/SnakeGame.py
@@ -155,9 +155,9 @@ class SnakeGame:
 
         ## 6. Set a negative reward if the snake head is adjacent to the snake body.
         # This is to discourage snake collisions.
-        for segment in self.snake[1:]:
-            if abs(self.head.x - segment.x) < 2 and abs(self.head.y - segment.y) < 2:
-                reward -= -2
+        # for segment in self.snake[1:]:
+        #    if abs(self.head.x - segment.x) < 2 and abs(self.head.y - segment.y) < 2:
+        #        reward -= -2
 
         self.game_reward += reward
         self.game_board.update_snake(snake=self.snake, direction=self.direction)